+++ src/Category_Colors/Extras.php
@@ -35,6 +35,8 @@ class Extras extends Frontend {
 		$css   = array();
 		$css[] = "#tribe-events-content div.tribe-events-category-{$slug}.hentry.vevent .tribe-events-tooltip h4.entry-title,";
 		$css[] = ".tribe-grid-body .tribe-events-week-hourly-single:hover.tribe-events-category-{$slug},";
+		$css[] = ".tribe-grid-body .tribe-events-week-hourly-single.tribe-events-category-{$slug},";
+		$css[] = ".tribe-grid-allday .tribe-events-week-allday-single.tribe-events-category-{$slug},";
 		$css[] = '';
 		$css   = implode( "\n", $css );
 		echo $css;