+++ views/optionsform.php
@@ -1,9 +1,6 @@
 <?php
 namespace Fragen\Category_Colors;
 
-$teccc            = Main::instance();
-$options          = Admin::fetch_options( $teccc );
-$teccc->all_terms = ! empty( $teccc->all_terms ) ? $teccc->all_terms : $options['all_terms'];
 $teccc->setup_terms( $options );
 
 ?>