+++ CHANGES.md
@@ -1,6 +1,7 @@
 #### [unreleased]
 * fixed widget CSS
 * move list CSS to `class Extras`
+* fixed week CSS
 
 #### 5.2.2 / 2018-07-28
 * fix to create CSS when files don't exist