+++ CHANGES.md
@@ -2,6 +2,7 @@
 * fixed widget CSS
 * move list CSS to `class Extras`
 * fixed week CSS
+* load stylesheet all the time, more efficient as external file
 
 #### 5.2.2 / 2018-07-28
 * fixed to create CSS when files don't exist