+++ CHANGES.md
@@ -7,6 +7,7 @@
 * use `site_url()` for stylesheet to accomodate  subdirectory install
 * refactor setup of ignored terms and term data
 * add `border-right` to featured events
+* updated validation code upon saving options
 
 #### 5.2.2 / 2018-07-28
 * fixed to create CSS when files don't exist