+++ CHANGES.md
@@ -4,10 +4,10 @@
 * fixed week CSS
 * load stylesheet all the time, more efficient as external file
 * updated `class Main` to `use Tribe__Events__Main` can now use `Tribe__Events__Main::TAXONOMY`
-* use `site_url()` for stylesheet to accomodate  subdirectory install
 * refactor setup of ignored terms and term data
 * add `border-right` to featured events
 * updated validation code upon saving options
+* use `wp_upload_url()` to accomodate user install preferences
 
 #### 5.2.2 / 2018-07-28
 * fixed to create CSS when files don't exist