+++ CHANGES.md
@@ -8,6 +8,7 @@
 * add `border-right` to featured events
 * updated validation code upon saving options
 * use `wp_upload_url()` to for stylesheet to accomodate user directory preferences
+* set SSL corrected URLs for `wp_upload_url()`, <https://core.trac.wordpress.org/ticket/25449>
 
 #### 5.2.2 / 2018-07-28
 * fixed to create CSS when files don't exist